+++ test/oauth/http/requesttokenstore.test.js
@@ -3,8 +3,8 @@
 var expect = require('chai').expect;
 var $require = require('proxyquire');
 var sinon = require('sinon');
-var factory = require('../../../com/oauth/http/statestore');
-var RequestTokenStore = require('../../../lib/oauth/statestore');
+var factory = require('../../../com/oauth/http/requesttokenstore');
+var RequestTokenStore = require('../../../lib/oauth/requesttokenstore');
 
 
 describe('oauth/http/requesttokenstore', function() {
@@ -16,8 +16,8 @@ describe('oauth/http/requesttokenstore', function() {
   
   it('should construct RequestTokenStore', function() {
     var RequestTokenStoreSpy = sinon.spy(RequestTokenStore);
-    var factory = $require('../../../com/oauth/http/statestore', {
-      '../../../lib/oauth/statestore': RequestTokenStoreSpy
+    var factory = $require('../../../com/oauth/http/requesttokenstore', {
+      '../../../lib/oauth/requesttokenstore': RequestTokenStoreSpy
     });
     
     var store = factory();